Move helper method to the pytest-pixl package, as per review comment
